Update AvgNew.h
Either header or source has int16_t percentile(int[16_t] pos) declared incorrectly

+++ libraries/AvgNew/AvgNew.h
@@ -87,7 +87,7 @@ class Average {
         uint16_t sum();
         void clear();
         void bubble_sort();
-        int16_t percentile(int pos);
+        int16_t percentile(int16_t pos);
         //Average<T> &operator=(Average<T> &a);
 
 };
